feat: migrate ProfileAvatar to TS
